add dedicated job runner
