+++ crates/kitsune-core/Cargo.toml
@@ -11,6 +11,10 @@ async-stream = "0.3.5"
 async-trait = "0.1.73"
 athena = { path = "../../lib/athena" }
 autometrics = { version = "0.6.0", default-features = false }
+aws-credential-types = { version = "0.56.1", features = [
+    "hardcoded-credentials",
+] }
+aws-sdk-s3 = "0.30.0"
 base64-simd = "0.8.0"
 bytes = "1.5.0"
 const_format = "0.2.31"
@@ -73,7 +77,9 @@ zxcvbn = { version = "2.2.2", default-features = false }
 
 [features]
 default = []
+kitsune-search = ["kitsune-search/kitsune-search"]
 mastodon-api = []
+meilisearch = ["kitsune-search/meilisearch"]
 
 [build-dependencies]
 vergen = { version = "8.2.5", features = ["build", "git", "gitcl"] }
