+++ crates/kitsune-core/build.rs
@@ -1,8 +1,6 @@
 use vergen::EmitBuilder;
 
 fn main() {
-    println!("cargo:rerun-if-changed=templates");
-
     EmitBuilder::builder()
         .all_git()
         .git_sha(true)
