+++ kitsune-job-runner/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "kitsune-job-runner"
+edition.workspace = true
+version.workspace = true
+
+[dependencies]
+athena = { path = "../lib/athena" }
+clap = { version = "4.4.3", features = ["derive"] }
+color-eyre = "0.6.2"
+deadpool-redis = "0.12.0"
+just-retry = { path = "../lib/just-retry" }
+kitsune-core = { path = "../crates/kitsune-core" }
+kitsune-db = { path = "../crates/kitsune-db" }
+tokio = { version = "1.32.0", features = ["full"] }
+toml = "0.8.0"
+tracing = "0.1.37"
+
+[features]
