+++ kitsune/src/main.rs
@@ -3,8 +3,8 @@
 
 use color_eyre::{config::HookBuilder, Help};
 use eyre::Context;
-use kitsune::{consts::STARTUP_FIGLET, http};
-use kitsune_core::{config::Configuration, consts::VERSION, job};
+use kitsune::consts::STARTUP_FIGLET;
+use kitsune_core::{config::Configuration, consts::VERSION};
 use std::{
     borrow::Cow,
     env, future,
@@ -147,12 +147,12 @@ async fn boot() -> eyre::Result<()> {
     .context("Failed to connect to and migrate the database")
     .with_suggestion(|| postgres_url_diagnostics(&config.database.url))?;
 
-    let job_queue = kitsune::prepare_job_queue(conn.clone(), &config.job_queue)
+    let job_queue = kitsune_job_runner::prepare_job_queue(conn.clone(), &config.job_queue)
         .context("Failed to connect to the Redis instance for the job scheduler")?;
     let state = kitsune::initialise_state(&config, conn, job_queue.clone()).await?;
 
-    tokio::spawn(self::http::run(state.clone(), config.server.clone()));
-    tokio::spawn(self::job::run_dispatcher(
+    tokio::spawn(kitsune::http::run(state.clone(), config.server.clone()));
+    tokio::spawn(kitsune_job_runner::run_dispatcher(
         job_queue,
         state.core.clone(),
         config.job_queue.num_workers.get(),
